Fixes addressing feedback
